demucs_requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
demucs @ git+https://github.com/facebookresearch/demucs@e976d93ecc3865e5757426930257e200846a520a
dm-tree==0.1.9
docopt==0.6.2
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.30.2
HyperPyYAML==1.2.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
lightning==2.5.1
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
namex==0.0.8
networkx==3.4.2
numba==0.61.2
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
openunmix==1.3.0
opt_einsum==3.4.0
optree==0.15.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
pooch==1.8.2
primePy==1.3
propcache==0.3.1
protobuf==4.25.6
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
requests==2.28.1
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.40
starlette==0.46.1
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
threadpoolctl==3.6.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.15.2+cpu
tqdm==4.67.1
treetable==0.2.5
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0